Adds JSON-js cycle fix to the %j option for circular references
Objects with circular references do not work with the %j option.  This change adds
Douglas Crockfords JSON cycle code to the arg parameter as a wrapper around the
input.  If there is a circular reference in the object this will short circuit the
duplicates and replace them with a reference string (removing dups and short
circuiting the circular reference).

https://github.com/douglascrockford/JSON-js